+++ src/components/bluepad32/uni_esp32.c
@@ -42,39 +42,12 @@ limitations under the License.
 _Static_assert(CONFIG_BTDM_CTRL_BR_EDR_MAX_ACL_CONN >= 2, "Max ACL must be >= 2");
 #endif  // UNI_ENABLE_BREDR
 
-// This function should be called from "app_main"
-int uni_esp32_main(void) {
-    // hci_dump_open(NULL, HCI_DUMP_STDOUT);
-
-#ifdef CONFIG_BLUEPAD32_UART_OUTPUT_ENABLE
-    uni_esp32_enable_uart_output(1);
-#else
-    // Adafruit Airlift modules have the UART RX/TX (GPIO 1 / 3) wired with the
-    // controller so they can't be used for logging. In fact they can generate
-    // noise and can break the communication with the controller.
-    uni_esp32_enable_uart_output(0);
-#endif
-
-    // Configure BTstack for ESP32 VHCI Controller
-    btstack_init();
-
-    // hci_dump_init(hci_dump_embedded_stdout_get_instance());
-
-    // Init Bluepad32.
-    uni_main(0 /* argc */, NULL /* argv */);
-
-    // Does not return.
-    btstack_run_loop_execute();
-
-    return 0;
-}
-
 // Code taken from nina-fw
 // https://github.com/adafruit/nina-fw/blob/master/main/sketch.ino.cpp
 //
 // Bluepad32 is compiled with UART RX/TX disabled by default.
 // But can be enabled/disabled in runtime by calling this function.
-void uni_esp32_enable_uart_output(int enabled) {
+void uni_esp32_enable_uart_output(bool enabled) {
 #ifdef CONFIG_IDF_TARGET_ESP32
     if (enabled) {
         PIN_FUNC_SELECT(GPIO_PIN_MUX_REG[1], 0);
@@ -102,4 +75,15 @@ void uni_esp32_enable_uart_output(int enabled) {
         ets_install_putc2(NULL);
     }
 #endif  // CONFIG_IDF_TARGET_ESP32
+}
+
+void uni_esp32_init(void) {
+#ifdef CONFIG_BLUEPAD32_UART_OUTPUT_ENABLE
+    uni_esp32_enable_uart_output(1);
+#else
+    // Adafruit Airlift modules have the UART RX/TX (GPIO 1 / 3) wired with the
+    // controller so they can't be used for logging. In fact they can generate
+    // noise and can break the communication with the controller.
+    uni_esp32_enable_uart_output(0);
+#endif
 }