+++ src/main/main.c
@@ -15,17 +15,43 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 ****************************************************************************/
+#include <stdlib.h>
+
+#include <btstack_port_esp32.h>
+#include <btstack_run_loop.h>
 
 #include "sdkconfig.h"
+#include "uni_main.h"
+#include "uni_platform.h"
+
 #ifndef CONFIG_BLUEPAD32_PLATFORM_ARDUINO
 
 // Not needed in Arduino platform since it has "autostart" code defined in
 // uni_platform_arduino.
 
-#include "uni_esp32.h"
+// Defined in my_platform.c
+struct uni_platform* get_my_platform(void);
 
 int app_main(void) {
-    return uni_esp32_main();
+    // hci_dump_open(NULL, HCI_DUMP_STDOUT);
+
+    // Configure BTstack for ESP32 VHCI Controller
+    btstack_init();
+
+    // hci_dump_init(hci_dump_embedded_stdout_get_instance());
+
+#ifdef CONFIG_BLUEPAD32_PLATFORM_CUSTOM
+    // Must be called before uni_main()
+    uni_platform_set_custom(get_my_platform());
+#endif  // CONFIG_BLUEPAD32_PLATFORM_CUSTOM
+
+    // Init Bluepad32.
+    uni_main(0 /* argc */, NULL /* argv */);
+
+    // Does not return.
+    btstack_run_loop_execute();
+
+    return 0;
 }
 
 #endif  //  CONFIG_BLUEPAD32_PLATFORM_ARDUINO