+++ src/main/my_platform.c
@@ -213,7 +213,7 @@ static void trigger_event_on_gamepad(uni_hid_device_t* d) {
 //
 // Entry Point
 //
-struct uni_platform* my_platform_create(void) {
+struct uni_platform* get_my_platform(void) {
     static struct uni_platform plat = {
         .name = "custom",
         .init = my_platform_init,